Always apply the datadog android gradle plugin
